\name{coef.stbss}
\alias{coef.stbss}

\title{
Coef Method for an Object of Class 'stbss'
}

\description{
Extracts the estimated unmixing matrix of an object of class \code{'stbss'}.
}

\usage{
\method{coef}{stbss}(object, ...)
}

\arguments{
  \item{object}{
object of class \code{'stbss'}. Usually result of \code{\link{stbss}}.
}
  \item{\dots}{
further arguments to be passed to or from methods.
}
}

\value{
Returns the estimated unmixing matrix of an object of class \code{'stbss'} as a numeric matrix. For a description of the class \code{'stbss'} see \code{\link{stbss}}.
}

\seealso{
\code{\link{stbss}}
}
