\name{print.stbss}
\alias{print.stbss}

\title{
Print Method for an Object of Class 'stbss'
}

\description{
Prints the estimated unmixing matrix, the (pseudo-)eigenvalues and the diagonalized local autocovariance matrices for an object of class \code{'stbss'}.
}

\usage{
\method{print}{stbss}(x, \dots)
}

\arguments{
  \item{x}{
object of class \code{'stbss'}. Usually result of \code{\link{stbss}}.
}
  \item{\dots}{
additional arguments for the method \code{print.listof}.
}
}

\value{
No return value.
} 

\seealso{
\code{\link{stbss}}
}

